Update gas phase RISM parameters again
Turns out we should not have loose 1d convergence thresholds. It also
turns out that when you don't specify mdiis1d_step, RISM's default value
of 0.5 gets overriden to a value of 0.1 if the number of sites in the
solvent is less than or equal to 6. But if we manually set the step to
0.5, it overrides this logic and falls back to 0.5, which is a bad step
for molecular calculations.