Start using scaled_positions in `constrain_slab`
If we use absolute positions, then we end up incorrectly constraining
slabs whose unit cells are tilted with respect to the Z-direction. This
commit reverts some features of an older version of `constrain_slab`
where we atomic scaled positions that are scaled with respect to the
lattice vectors. See #89 for more details.