Try new way of moving QE initialization files
We used a `ScriptTask` with a `mv *` command before, but for some reason
the `mv` command  was happening *after* subsequent FireTasks in the
FireWork (i.e., it ran RISM first and then moved files, but we want it
the other way around). Maybe FireWorks' built-in `FileTransferTask` will
do it more correctly.