Tentatively stop trying to seed RISM with QE
We keep getting "error opening xml data file" errors during the first
RISM scf. Did a minor test where we took out the `startingpot` and
`startingwfc` flags, and it started running just fine. Going to take
those default values out so we effectively ignore the 1-step QE seed for
RISM. Leaving the rest of the framework in until we figure out more of
what's going on.