Undo some more of the pre-initialization for RISM
The timing of the moving of files is all messed up. We end up doing the
RISM relaxation before the move is done, and so the first part of the
RISM output is gone, which we need to read the output file. It also
results in a bunch of binary at the beginning of the output file. Doing
this move thing is just too dirty, so we're going to punt on the
pre-initialization for now.